requirements.txt
# Core dependencies
click>=8.1.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
python-docx>=0.8.11

# Embeddings and vector operations
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0

# Knowledge Graph (Neo4j)
neo4j>=5.12.0
py2neo>=2021.2.3

# NLP and entity extraction
spacy>=3.6.0
transformers>=4.30.0
torch>=2.0.0

# LLM interaction
requests>=2.31.0
httpx>=0.24.0

# Evaluation metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.0

# Data handling
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-html>=3.2.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
